+++ src/gretapy/pl/__init__.py
@@ -1,3 +1,4 @@
+from ._distance import cre_to_tss_distance
 from ._heatmap import heatmap
 from ._links import links
 from ._ranking import ranking